chore: move RealmTag sync operations to TagsRepository
- Moves `insert` and `getTagsArray` out of `RealmTag` companion object to enforce cleaner architecture.
- Implements `insertFromJson` and `getTagsArray` within `TagsRepositoryImpl`.
- Injects `TagsRepository` into `TransactionSyncManager` via `ServiceModule`.
- Replaces static call sites in `TransactionSyncManager`, `CoursesRepositoryImpl`, and `ResourcesRepositoryImpl` with new repository instance methods.
- Updates `RealmTag.setAttachedTo` visibility to public to allow proper managed `RealmList` mutation from the repository.
