+++ app/src/main/java/org/ole/planet/myplanet/repository/TagsRepositoryImpl.kt
@@ -1,8 +1,11 @@
 package org.ole.planet.myplanet.repository
 
+import com.google.gson.JsonArray
+import com.google.gson.JsonObject
 import javax.inject.Inject
 import org.ole.planet.myplanet.data.DatabaseService
 import org.ole.planet.myplanet.model.RealmTag
+import org.ole.planet.myplanet.utils.JsonUtils
 
 class TagsRepositoryImpl @Inject constructor(
     databaseService: DatabaseService
@@ -102,5 +105,36 @@ class TagsRepositoryImpl @Inject constructor(
 
         val parentsById = parents.associateBy { it.id }
         return tagIds.mapNotNull { parentsById[it] }
+    }
+
+    override fun insertFromJson(mRealm: io.realm.Realm, act: JsonObject) {
+        var tag = mRealm.where(RealmTag::class.java).equalTo("_id", JsonUtils.getString("_id", act)).findFirst()
+        if (tag == null) {
+            tag = mRealm.createObject(RealmTag::class.java, JsonUtils.getString("_id", act))
+        }
+        if (tag != null) {
+            tag._rev = JsonUtils.getString("_rev", act)
+            tag._id = JsonUtils.getString("_id", act)
+            tag.name = JsonUtils.getString("name", act)
+            tag.db = JsonUtils.getString("db", act)
+            tag.docType = JsonUtils.getString("docType", act)
+            tag.tagId = JsonUtils.getString("tagId", act)
+            tag.linkId = JsonUtils.getString("linkId", act)
+            val el = act["attachedTo"]
+            if (el != null && el.isJsonArray) {
+                tag.setAttachedTo(JsonUtils.getJsonArray("attachedTo", act))
+            } else {
+                tag.attachedTo?.add(JsonUtils.getString("attachedTo", act))
+            }
+            tag.isAttached = (tag.attachedTo?.size ?: 0) > 0
+        }
+    }
+
+    override fun getTagsArray(tags: List<RealmTag>): JsonArray {
+        val array = JsonArray()
+        for (t in tags) {
+            array.add(t._id)
+        }
+        return array
     }
 }
